refactor: rename next-plugin to actual package name
