+++ .travis.yml
@@ -9,7 +9,7 @@ node_js:
   - "11.10"
 install:
   - npm install
-  - cd packages/next-plugin && npm install && cd ../../
+  - cd packages/serverless-nextjs-plugin && npm install && cd ../../
   - cd integration/app-with-serverless-offline && npm install && cd ../../
 script:
   - npm run lint
