+++ packages/serverless-nextjs-plugin/classes/NextPage.js
