+++ packages/serverless-nextjs-plugin/classes/PluginBuildDir.js
