+++ packages/serverless-nextjs-plugin/classes/__tests__/NextPage.test.js
