+++ packages/serverless-nextjs-plugin/classes/__tests__/PluginBuildDir.test.js
