+++ packages/serverless-nextjs-plugin/index.js
