+++ packages/serverless-nextjs-plugin/lib/__tests__/addCustomStackResources.test.js
