+++ packages/serverless-nextjs-plugin/lib/__tests__/build.test.js
