+++ packages/serverless-nextjs-plugin/lib/__tests__/copyBuildFiles.test.js
