+++ packages/serverless-nextjs-plugin/lib/__tests__/displayStackOutput.test.js
