+++ packages/serverless-nextjs-plugin/lib/__tests__/getAssetsBucketName.test.js
