+++ packages/serverless-nextjs-plugin/lib/__tests__/getFactoryHandlerCode.test.js
