+++ packages/serverless-nextjs-plugin/lib/__tests__/parseNextConfiguration.test.js
