+++ packages/serverless-nextjs-plugin/lib/__tests__/rewritePageHandlers.js
