+++ packages/serverless-nextjs-plugin/lib/__tests__/uploadStaticAssets.test.js
