+++ packages/serverless-nextjs-plugin/lib/addCustomStackResources.js
