+++ packages/serverless-nextjs-plugin/lib/build.js
