+++ packages/serverless-nextjs-plugin/lib/copyBuildFiles.js
