+++ packages/serverless-nextjs-plugin/lib/displayStackOutput.js
