+++ packages/serverless-nextjs-plugin/lib/getAssetsBucketName.js
