+++ packages/serverless-nextjs-plugin/lib/getFactoryHandlerCode.js
