+++ packages/serverless-nextjs-plugin/lib/getNextPagesFromBuildDir.js
