+++ packages/serverless-nextjs-plugin/lib/parseNextConfiguration.js
