+++ packages/serverless-nextjs-plugin/lib/rewritePageHandlers.js
