+++ packages/serverless-nextjs-plugin/lib/uploadStaticAssets.js
