+++ packages/serverless-nextjs-plugin/package.json
