+++ packages/serverless-nextjs-plugin/resources/api-gw-proxy.yml
