+++ packages/serverless-nextjs-plugin/resources/assets-bucket.yml
