+++ packages/serverless-nextjs-plugin/utils/logger.js
