+++ packages/serverless-nextjs-plugin/utils/pathToPosix.js
