+++ packages/serverless-nextjs-plugin/utils/s3/upload.js
