+++ packages/serverless-nextjs-plugin/utils/test/parsedNextConfigurationFactory.js
